Pre-work (part 1/3): Remove clutter
Laravel comes with an initial blade template ready to go (Welcome.blade.php), which has lots of elements to it (Head style tag, basic login routing, paragraphs of comments).

This commit removes all unnecessary bits. The focus of this project is to show how to set up Vue-Typescript on Laravel. Therefore, the only thing I want the Welcome template to have is the instantiation of the ExampleComponent.vue that laravel brings with it. (On the next commit we'll make that possible)