Pre-work (part 2/3): Wire-up regular vue
Laravel comes with an example vue component ready to go, but the starting Blade template does not reference the compiled javascript (nor css) file at all. This commit will add those references to the Welcome.blade.php template.

Another important point is that the global axios object tries to fetch the csrf_token from a meta tag called 'csrf-token'. However, this tag is not initially present on the Welcome.blade.php template either. So this commit also adds that.