Part 4 - Allow vue components to be written in Typescript
To have vue components working with Typescript, we just need to use the vue-property-decorator library.
In this commit I have also added a few features to demonstrate how to properly use Typescript with vue.

At this point, we have the entire front-end written in Typescript. In the next commit, we'll set up jest to test our Vue components.