Part 5 - Set up test suite for Typescript Vue components
Jest is a very popular and reliable test framework for javascript projects. It can work with Vue components with the help of vue-jest, and with Typescript files with the help of ts-jest.

Vue-test-utils offers a series of tools to interact with and retrieve information from Vue Components.

You may notice all these libraries were added in the package.json, as well as the configuration necessary for Jest to know where to find the test files.